move models to subpackage users, rooms or reservations
