+++ server/src/main/kotlin/org/tod87et/roomkn/server/models/rooms/NewRoomInfo.kt
@@ -1,10 +1,9 @@
-package org.tod87et.roomkn.server.models
+package org.tod87et.roomkn.server.models.rooms
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RoomInfo(
-    val id: Int,
+data class NewRoomInfo(
     val name: String,
     val description: String,
 )
