+++ server/src/main/kotlin/org/tod87et/roomkn/server/models/rooms/RoomInfo.kt
@@ -1,9 +1,10 @@
-package org.tod87et.roomkn.server.models
+package org.tod87et.roomkn.server.models.rooms
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class NewRoomInfo(
+data class RoomInfo(
+    val id: Int,
     val name: String,
     val description: String,
 )
