+++ server/src/main/kotlin/org/tod87et/roomkn/server/models/users/ShortUserInfo.kt
@@ -1,10 +1,9 @@
-package org.tod87et.roomkn.server.models
+package org.tod87et.roomkn.server.models.users
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserInfo(
+data class ShortUserInfo(
     val id: Int,
     val username: String,
-    val email: String,
 )
