+++ server/src/main/kotlin/org/tod87et/roomkn/server/models/users/UserInfo.kt
@@ -1,9 +1,10 @@
-package org.tod87et.roomkn.server.models
+package org.tod87et.roomkn.server.models.users
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ShortUserInfo(
+data class UserInfo(
     val id: Int,
     val username: String,
+    val email: String,
 )
