+++ server/src/test/kotlin/org/tod87et/roomkn/server/AuthTest.kt
@@ -13,8 +13,8 @@ import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.tod87et.roomkn.server.KtorTestEnv.testJsonApplication
 import org.tod87et.roomkn.server.database.DatabaseFactory
-import org.tod87et.roomkn.server.models.LoginUserInfo
-import org.tod87et.roomkn.server.models.UnregisteredUserInfo
+import org.tod87et.roomkn.server.models.users.LoginUserInfo
+import org.tod87et.roomkn.server.models.users.UnregisteredUserInfo
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
