+++ server/src/test/kotlin/org/tod87et/roomkn/server/DatabaseTest.kt
@@ -9,14 +9,14 @@ import org.tod87et.roomkn.server.database.ConstraintViolationException
 import org.tod87et.roomkn.server.database.DatabaseSession
 import org.tod87et.roomkn.server.database.MissingElementException
 import org.tod87et.roomkn.server.database.ReservationException
-import org.tod87et.roomkn.server.models.NewRoomInfo
-import org.tod87et.roomkn.server.models.RegistrationUserInfo
-import org.tod87et.roomkn.server.models.Reservation
-import org.tod87et.roomkn.server.models.RoomInfo
-import org.tod87et.roomkn.server.models.ShortRoomInfo
-import org.tod87et.roomkn.server.models.ShortUserInfo
-import org.tod87et.roomkn.server.models.UnregisteredReservation
-import org.tod87et.roomkn.server.models.UserInfo
+import org.tod87et.roomkn.server.models.rooms.NewRoomInfo
+import org.tod87et.roomkn.server.models.users.RegistrationUserInfo
+import org.tod87et.roomkn.server.models.reservations.Reservation
+import org.tod87et.roomkn.server.models.rooms.RoomInfo
+import org.tod87et.roomkn.server.models.rooms.ShortRoomInfo
+import org.tod87et.roomkn.server.models.users.ShortUserInfo
+import org.tod87et.roomkn.server.models.reservations.UnregisteredReservation
+import org.tod87et.roomkn.server.models.users.UserInfo
 import javax.sql.DataSource
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
